Added ordered handling of markup specials
Previously markup specials were stored in a Hash.

Depending on hash traversal order on Ruby 1.8 specials could be
processed in a different order.  This could lead to bugs in markup
processing across runs.

Markup specials are now stored in an Array.  This allows specials to be
processed in the order they were inserted on all ruby versions.

Additionally, RDoc::RDoc#document now calls self.store= to create a
connection between the driver and the store for HTML generation.