Don't delete attribute data when re-saving a class
The RDoc store places attributes and instance methods in the same place
in the cache.  These are stored in the same place to make method lookup
with ri easier.

To determine which ri data files to delete from the store when saving an
update to a class ri subtracted the cached instance method names from
the instance method list but not the attribute names.

This caused RDoc to delete attribute data that was still present.

Now RDoc subtracts both the instance methods and attributes from the
cache list.  This prevents accidental deletion of ri data.

Fixes #171