Raise an Error on a unterminated heredoc
When processing verbatim blocks RDoc attempts to detect if they are ruby
code.  If the verbatim block contains an unterminated here-document a
NoMethodError would be raised.  This causes RDoc to crash.

Now an RDoc::RubyLex::Error is raised which allows the HTML processor to
create an un-highlighted <pre> element instead.

Reported by darix