Added test for #180
When loading a class or module from the RDoc store then reparsing its
source file the old comment is replaced with the new comment.

This also means that if a module or class is mentioned multiple times in
the same file RDoc will only use the last comment.

These two behaviors of the comment conflict, so I must delay fixing this
issue until the future while I think about it.