Separate C call-seq for "copied" methods
If the same C function is used to implement two methods typically the
call-seq in the comment describes both methods together.

This leads to confusing duplication in generated HTML as identical
content is show twice.

Now when such a method is encountered the call-seq is de-duplicated to
avoid this confusion.