Reset full_name after duping a CodeObject
When an existing method is converted to a module_function RDoc would dup
the object including the full_name.  RDoc::Store would parse the
full_name to determine how to save the file which would result in the
method appearing only as an instance method.

Now the full_name is cleared when a CodeObject is duped allowing the
correct name to be used in RDoc::Store.

Part of #222 and Ruby bug #8225