Fix deduplication of C methods with ==
C methods can share the same C function and comment.  When deduplicating
the call-seq methods that did not start with a letter were incorrectly
deduplicated due to special handling for [] and []=.  Additionally, the
selection of operator-style methods was insufficient.

Now special handling for [] and []= is only performed for those methods.

Now an extra check is made for operator-style methods.

Fixes #244