Stop adding "http://" to schema-less URLs
Since the Dawn of Time RDoc has added http:// to schema-less URLs.  I
can't think of a reason to support this, so I have removed this from the
formatter.

If this breaks compatibility, instead of reverting this commit, RDoc
format parser should be changed to add http:// to schema-less URLs.

Fixes #208