Support ruby 2.0 hash syntax for keywords
The lexer did not support the new hash syntax for keywords so a line
like:

    foo module: whatever

Would make the Ruby parser believe that a module named ":" was declared.
This would lead to an error on Windows as ":" is not allowed in a
filename.

Now the lexer creates a TkIDENTIFIER token for keywords followed by ":".

Fixes #245