Prevent \\<tag> from escaping the tag
Previously any number of \ before a tag (protectable sequence) would
cause the tag to be escaped.  This prevented users from placing a \
inside tags.

Now a tag can only be escaped with a single \.  \\<tag> does not escape
the tag, nor does \\\<tag> (which is arguably a new bug).

Fixes #251