Hard link template files to their destinations
Previously RDoc copied the template files to their destinations.  This
meant if there were a bug in RDoc that required a security patch (like
CVE-2013-0256) each file needs to be updated.

Now, if the OS and filesystem support it, the files that ship with the
RDoc template will be hard-linked.  If the OS are filesystem does not
support it the file will be copied, as before.

Fixes #186.