Use get_end_token in skip_optional_do_...
skip_optional_do_after_expression looks for an end_token the same as
method_parameters_end_token so method_parameters_end_token was renamed
to get_end_token and used in skip_optional_do_after_expression.

This also adds TkRPAREN handling when skipping the do which may cause
future issues (as the nil is not handled).