Fix rb_file_const() when used outside file.c
Previously rdoc assumed rb_file_const was called within file.c, but this
is not true in trunk.  This caused the FNM_* constants to go missing.

Now these constants appear in File::Constants as expected.

(The constants do not appear in File even though File::Constants is
included there, this is a limitation of other parts of RDoc.)