Ignore "]" following an RDOCLINK
This caused problems when an RDOCLINK was discovered inside a TIDYLINK.
The ] wouldn't indicate the end of the RDOCLINK (any non-whitespace) so
it would consume the ] and any following punctuation.

This would cause a TIDYLINK to be unrecognized.

The handling of markup inside a paragraph needs its own parser one of
these days.