Capitalize "Ruby" in comments
A user came through and capitalized all instances of "ruby" as that's
not the proper way to do things.  Since the commit came from the ruby
repository it did not update the kpeg files (which are not included
there).

If the literals parsers are rebuilt in the future, now the "Ruby" fix
won't regress.