Fix Document-attr in the C parser
Previously the comment-open marker was not included in the comment
returned from find_attr_comment.  This meant that normalization would
not strip the "*"s correctly leading to broken output.

Now the opening "*"s are included, but only when the Document-attr is on
the same line as the opening "/*" or on the line following the opening
"/*".

Fixes #271