Add hash equality to Section
When an item is not in any section it ends up in the nil section.
Unfortunately RDoc creates a separate Section object between grouping
and display time which would cause items to disappear from a section at
HTML generation time.

Now that two Section objects with the same title map to the same Hash
key the constants will not disappear.

Fixes bug #399