Support did_you_mean
DidYouMean::SpellChecker exists

Ruby 2.3 or later, did_you_mean gem is adopted as standard library. The
did_you_mean gem can be used as a library via DidYouMean::SpellChecker
at 1.0.3 or later, so checks that the class is defined and uses it. And
then, this feature will become enabled if the gem 1.0.3 or later is
installed.