Fix for warnings on Ruby 2.5 in ERB binding
ERB requires variable context as Binding object and Darkfish uses
Kernel#binding for it, but it causes warnings "assigned but unused
variable" on Ruby 1.9 or later, so there ware "suppress 1.9.3 warning"
code like below:

  # suppress 1.9.3 warning
  variable = variable = expressions

The "variable" variable is assigned but unused *at this scope* and used
inside ERB template, so the double assign of "variable" is for suppress
the warning.

This code causes warning "assigned but unused variable" on Ruby 2.5
again. I fixed it with Binding#local_variable_set for "use" variable.