Applied 1a629de from peg-markdown
Disallow blank space between [label] and ( in inline link.

The official markdown syntax description says that a blank
space is allowed between ] and [ in a reference style link,
but it does not say this for an inline link; instead, it says
that the ( must be immediately after the ].  This interpretation
is confirmed by the behavior of Markdown.pl.

This change brings peg-markdown into agreement with Markdown.pl,
PHP markdown, and pandoc.  Note that some markdown implementations
(discount, sundown) do allow the space.  Though something can be
said in favor of allowing the space, it seems best to go with
the spec to resolve such disagreements.

In addition, with short reference style links, it is easy to
do something like

    [my link] (word)

    [my link]: /url

which we wouldn't want to parse as an inline link.