Bypass ERB's Ruby syntax error
The legacy lexical analyzer and parser of RDoc 5 process ERB file with
no errors because of complex bug. RDoc 6 what uses Ripper causes errors
when processes ERB file. It was just bug fix but RDoc lost
compatibility. For example, when installs ActiveRecord 5.1.4 by
"gem install activerecord -v 5.1.4", it uses RDoc internal, so the RDoc
crashes if it's RDoc 6.0.0.

This commit provides compatibility behavior with RDoc 5 or older. When
an exception occurs inside Ruby parser, confirms that the source code
contains "<%" and "%>" to detect ERB file. Skips it if RDoc guesses that
the file is ERB by the logic.