Fix random NameError of RipperStateLex
Because RDoc::RipperStatelex is loaded by RDoc::Parser, some tests that
need RDoc::RipperStateLex fail when RDoc::Parser isn't autoloaded. That
whether RDoc::RipperStateLex is loaded or not depends on RDoc::Parser is
autoloaded. Therefore NameError of RDoc::RipperStateLex is occurred at
random.

This patch fixes it with that RDoc::RipperStateLex is moved to
RDoc::Parser::RipperStateLex.