Use original method notation
At 297cd5a2, method names are replaced with normalized text. But the
feature has critical problems:

1) Some methods are replaced with parents class receiver

For example, the document of BasicSocket describes #read_nonblock for
inheritance, but it's replaced with IO#read_nonblock. This behavior
confuses users.

2) Arguments are removed

For example, the document of Time describes like below:

  Time.utc(99) means the year 99 AD, not 1999 AD.

But it's replaced with:

  Time.utc means the year 99 AD, not 1999 AD.

This behavior clearly is a bug.

This patch removes the normalization logic.