Disable wrap processing in paragraph
I'm using Stackprof for a performance check of Rails documents
generation.

Before this commit, RDoc::Text#wrap uses processing time about 12.7%.

  ~/rdoc$ stackprof stackprof-rails.dump --text --limit 3:
  ==================================
    Mode: cpu(1000)
    Samples: 94637 (1.54% miss rate)
    GC: 18742 (19.80%)
  ==================================
  TOTAL   (pct) SAMPLES   (pct) FRAME
  18742 (19.8%)   18742 (19.8%) (garbage collection)
  12022 (12.7%)   12022 (12.7%) RDoc::Text#wrap
  17468 (18.5%)    6836  (7.2%) RDoc::Generator::Darkfish#generate_class

After this commit, processing time of RDoc::Text#wrap is reduced to
0.0%.

  $ stackprof stackprof-rails.dump --method 'RDoc::Text#wrap'
  RDoc::Text#wrap (/home/aycabta/rdoc/lib/rdoc/text.rb:269)
    samples:     9 self (0.0%)  /      9 total (0.0%)
    callers:

The wrap processing is just for readability of HTML, but it's too heavy
and unnecessary in the present day.