Fix full name of known class
Properly set the name of `File::Constants`, which is the only name
with a namespace in `RDoc::KNOWN_CLASSES`.  When it is generated
by `rb_file_const` in dir.c, `name` is set to the qualified name
as same as `full_name`, and generated in the normal way in file.c
later, already set `full_name` is cleared and `name` will be
constructed from the enclosing namespace and the `name`.  It will
results in duplicated namespace, `File::File::Constants`.