Update rake ghpages task
If the GitHub pages source is set to the "gh-pages" branch and the
"/doc" folder, then this should get everything back up-to-date.

The older version should probably work, too.  But I never like running
`rm -rf *` if I can avoid it!  I think the other changes are minor
improvements, too.

It should be possible to run this rake task from a GitHub action.