Fix section comments in darkfish
This is almost certainly the wrong way to do this.  But the lack of
section comments has been bugging me for years, and this was my first
attempt at figuring out why it is broken.  Maybe someone else can fix it
The Right Way, or give me some advice on how to do so myself...