test/rdoc/test_rdoc_generator_json_index.rb: pend in test_generate in ppc64le.
We observed that this test randomly fails in the ruby/ruby Travis ppc64le case.
This commit is to pend the test_generate if the assertion for the generated
file's modified time fails in a ppc64le environment.

Note that I didn't use the word "Travis CI" or Travis CI specific environment variables
such as `TRAVIS` and `TRAVIS_CPU_ARCH`[1] in the code. Because I wanted to prioritize the
rdoc's independence from the ruby/ruby.

[1] https://docs.travis-ci.com/user/environment-variables/#default-environment-variables