Mention `.document` file in README
Rdoc has `.document` file in its root directory, allowing us to
determine which files are the target for documentation.
However, this fact is not mentioned anywhere in the doc including
README, which makes it quite difficult to utilize this feature.
This commit adds a mention to `.document` file in README so that
users can find the feature more easily.

One thing I'm not sure is if it's the right place.
Currently README includes some usage information, which can be a
separate file.