Update ruby-core workflow to just generate documentation
Since RDoc is now a bundled gem, we don't need to run test with Ruby core
with it commit by commit. But it'd still be beneficial to generate
documentation with the current RDoc commit and the Ruby master branch.

So this PR updates the Ruby core workflow to just do that.