Add a `build:local_ruby` task to simplify local testing flow
A common way to test RDoc changes is to build `ruby/ruby`'s documentation
with the latest RDoc changes. When RDoc was a default gem, we can
sync it to `ruby/ruby` with its `tool/sync_default_gems.rb` script.

Now that RDoc is a bundled gem, we need to use a different method to
sync it to `ruby/ruby`. And so far building it and moving it to
`ruby/ruby`'s bundled gems folder is the easiest way to do it.