Remove indirection from RDoc's call_seq extraction
There is no need to pass the method object to RDoc::Comment#extract_call_seq
just to assign the result to the method object's call_seq attribute. This
unnecessarily complicates the code and makes it harder to understand/refactor,
which is evident in the method's tests.

This commit removes the indirection and simplifies the implementation &
tests.