Avoid using block binding to capture template variables
Under the current implementation, the template variables are captured
through the `render_template` method's block. This makes it hard to
understand if a variable is intended to be used in the template or not.

This commit changes the `render_template` method to require passing
a hash of local variables to the template. This makes the variable
scope explicit and easier to reason about.