Centralize generator setup
Currently we need to call `setup` in individual generation methods,
despite having already called it in `RDoc#generate`.

This is because `setup` is lazily called when we run `ri`'s server
mode.

By calling `setup` eagerly in `Servlet#generator_for`, we can
avoid the repeated calls to `setup` in individual generation methods.