Image alt tag header formatting (#1320)
This allows us to parse the alt text from Markdown images, formatting a
header text correctly (using the alt text).

Without this PR, if a Markdown file has an image instead of a title,
that image will be used in the table of contents, which is unreadable
and undesirable.

Used in <https://github.com/rack/rack/pull/2313>.

---------

Co-authored-by: Stan Lo <stan001212@gmail.com>