Workaround `build` task issue in Ruby core CI
The issue is described in https://github.com/rubygems/rubygems/issues/8477

The issue fails the workflow whenever RDoc's version is bumped as there
will then be a mismatch between the target file name and the version
that's actually built.

For example:

```
rdoc 6.13.0 built to pkg/rdoc-6.12.0.gem.
```

This commit works around the issue by manually building the gem and
installing it locally so the latest installed version will match the
version contained in the gemspec file.