fix(options): `RDoc::Options` needs to call `parse` for its setup
This change workarounds the issue that `RDoc::Options.new` does not
work itself without calling `parse` since some important setup logic
lives in it.
I believe we can improve this a lot by moving default option setup
in `initialize`, for example, but it takes time.
This usage of `RDoc::Options.new` is documented in README, and
I believe this workaround solves it quickly and well enough.