Refactor PR preview workflow for improved approval process (#1370)
Because GitHub Actions' manual dispatching doesn't allow pre-filling the
PR number, it'd be tedious for maintainers to manually trigger the
workflow.

This refactor introduces a new workflow that checks for PRs from the
main repository or forks, and if it's a fork, it requires approval to
deploy. This works by associating the workflow with an environment that
requires approval to deploy.

The pr-review-check.yml workflow will then call the
cloudflare-preview.yml workflow to build and deploy the preview.