Allow rdoc-ref to link to non-text files (#1376)
It appears the current behavior of disallowing links to non-text files
is deliberate, as it was explicitly added in
3628e196cbbac4767f5b845c5b795c635a671b04. While the commit message
explains the change, it doesn't provide a justification for excluding
non-text TopLevels. The issue mentioned in the commit message is also
unrelated to the change.

It's just as useful to link to a non-text file as it is to link to a
text file, so I think it should be allowed. I was surprised when it
didn't work. I want to use this feature in tilt's documentation.

If there is a reason to disallow it by default, I think this limitation
should be documented. I also I think we should add an option to allow it
in that case.