Search automatically on ?q= URL query parameter (#1396)
Fixes https://github.com/ruby/rdoc/issues/1319 (I wanted this for
exactly that reason 😅)

Automatically start searching when ?q= parameter is given in the URL,
selecting and visiting the first match if available.

If not available, the query will be filled into the search box but we
won't navigate away.

This should enable configuring a generated rdoc page as a search engine
in one's web browser, or as a bang command in a search engine. For
reference Rails' sdoc (which I think rdoc's search is originally based
on) does this: https://api.rubyonrails.org/?q=changed%3F