Commit generated files (#1410)
Closes https://github.com/ruby/rdoc/issues/1107 and will likely prevent
issues like https://bugs.ruby-lang.org/issues/21541 in the future

With ignoring generated files in the source control, we always need one
extra step to make the gem usable, which limits how the gem can be
distributed (#1107) and make post-distribution issues easier to happen
(which are in general harder to fix too).

If gems like ~~`prism` and~~ `rbs` and `debug` can avoid this issue by
committing generated files and verifying them on CI, I think we should
do that too.