Change sidebar hiding strategy to avoid flickering on mobile devices (#1516)
Previously, the sidebar was hidden with JavaScript on small viewports.
This approach caused flickering when the page was loaded, especially on
mobile devices.

Now, the sidebar is hidden by default with the `hidden` attribute. On
large viewports, we display the sidebar with JavaScript. This approach
avoids flickering when the page is loaded, especially on mobile devices.